pyproject.toml
[project]
name = "github-issue-creator"
version = "0.1.0"
readme = "README.md"
authors = [
    {name = "Matous Havlena", email = "matous.havlena@ibm.com"}
]
license = {text = "Apache-2.0"}
requires-python = ">=3.11,<4.0"
dependencies = [
    "beeai-framework[agentstack]==0.1.61",
    "aiohttp>=3.13.2",
    "mcp>=1.20.0",
    "python-dotenv>=1.2.1",
    "openinference-instrumentation-beeai>=0.1.13",
    "agentstack-sdk>=0.4.0"
]

[project.optional-dependencies]
dev = [
    "watchfiles>=1.1.0",
    "ruff>=0.14.3"
]

[tool.ruff]
line-length = 120
target-version = "py311"

[project.scripts]
server = "github_issue_creator.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/github_issue_creator"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
